feat: make util for getting scoped actions
